backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
pytest
pytest-asyncio
pyyaml
jsonschema
python-dotenv